hjjoin
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@1290 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/unrar/Merge.java
@@ -65,9 +65,11 @@ public class Merge {
                 }
             });
             join.run();
-
+            if(autoDelete)
+            {
             for (int i = 0; i < files.length; i++) {
                 files[i].delete();
+            }
             }
             progress.finalize();
             return new File(join.sJoinedFile);
